requirements.txt
streamlit>=1.33.0
python-dotenv>=1.0.1
langchain>=0.1.20
langchain-openai>=0.1.7
langchain-community>=0.0.38
langchain-ollama>=0.1.0
pypdf>=4.2.0
tiktoken>=0.7.0
